requirements.txt
azure-functions
requests
beautifulsoup4
azure-cosmos